+++ extcommon/pod_count_check.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
-package extdeployment
+package extcommon
 
 import (
 	"context"
@@ -25,22 +25,24 @@ const (
 )
 
 type PodCountCheckAction struct {
+	Client                       *client.Client
+	ActionId                     string
+	TargetType                   string
+	SelectionTemplates           *action_kit_api.TargetSelectionTemplates
+	GetTarget                    func(request action_kit_api.PrepareActionRequestBody) string
+	GetDesiredAndCurrentPodCount func(k8s *client.Client, namespace string, target string) (*int32, int32, error)
 }
 
 type PodCountCheckState struct {
 	Timeout           time.Time
 	PodCountCheckMode string
 	Namespace         string
-	Deployment        string
+	Target            string
 	InitialCount      int
 }
 type PodCountCheckConfig struct {
 	Duration          int
 	PodCountCheckMode string
-}
-
-func NewPodCountCheckAction() action_kit_sdk.Action[PodCountCheckState] {
-	return PodCountCheckAction{}
 }
 
 var _ action_kit_sdk.Action[PodCountCheckState] = (*PodCountCheckAction)(nil)
@@ -52,7 +54,7 @@ func (f PodCountCheckAction) NewEmptyState() PodCountCheckState {
 
 func (f PodCountCheckAction) Describe() action_kit_api.ActionDescription {
 	return action_kit_api.ActionDescription{
-		Id:          PodCountCheckActionId,
+		Id:          f.ActionId,
 		Label:       "Pod Count",
 		Description: "Verify pod counts",
 		Version:     extbuild.GetSemverVersionStringOrUnknown(),
@@ -61,15 +63,9 @@ func (f PodCountCheckAction) Describe() action_kit_api.ActionDescription {
 		Kind:        action_kit_api.Check,
 		TimeControl: action_kit_api.TimeControlInternal,
 		TargetSelection: extutil.Ptr(action_kit_api.TargetSelection{
-			TargetType:          DeploymentTargetType,
+			TargetType:          f.TargetType,
 			QuantityRestriction: extutil.Ptr(action_kit_api.All),
-			SelectionTemplates: extutil.Ptr([]action_kit_api.TargetSelectionTemplate{
-				{
-					Label:       "default",
-					Description: extutil.Ptr("Find deployment by cluster, namespace and deployment"),
-					Query:       "k8s.cluster-name=\"\" AND k8s.namespace=\"\" AND k8s.deployment=\"\"",
-				},
-			}),
+			SelectionTemplates:  f.SelectionTemplates,
 		}),
 		Parameters: []action_kit_api.ActionParameter{
 			{
@@ -122,27 +118,26 @@ func (f PodCountCheckAction) Describe() action_kit_api.ActionDescription {
 }
 
 func (f PodCountCheckAction) Prepare(_ context.Context, state *PodCountCheckState, request action_kit_api.PrepareActionRequestBody) (*action_kit_api.PrepareResult, error) {
-	return preparePodCountCheckInternal(client.K8S, state, request)
-}
-
-func preparePodCountCheckInternal(k8s *client.Client, state *PodCountCheckState, request action_kit_api.PrepareActionRequestBody) (*action_kit_api.PrepareResult, error) {
 	var config PodCountCheckConfig
 	if err := extconversion.Convert(request.Config, &config); err != nil {
 		return nil, extension_kit.ToError("Failed to unmarshal the config.", err)
 	}
 
 	namespace := request.Target.Attributes["k8s.namespace"][0]
-	deployment := request.Target.Attributes["k8s.deployment"][0]
-	d := k8s.DeploymentByNamespaceAndName(namespace, deployment)
-	if d == nil {
-		return nil, extension_kit.ToError(fmt.Sprintf("Failed to find deployment %s/%s.", namespace, deployment), nil)
+	target := f.GetTarget(request)
+	desired, current, err := f.GetDesiredAndCurrentPodCount(f.Client, namespace, target)
+	if err != nil {
+		return nil, err
+	}
+	if desired == nil && (config.PodCountCheckMode == podCountEqualsDesiredCount || config.PodCountCheckMode == podCountLessThanDesiredCount) {
+		return nil, extension_kit.ToError(fmt.Sprintf("%s/%s has no desired count", namespace, target), nil)
 	}
 
 	state.Timeout = time.Now().Add(time.Millisecond * time.Duration(config.Duration))
 	state.PodCountCheckMode = config.PodCountCheckMode
 	state.Namespace = namespace
-	state.Deployment = deployment
-	state.InitialCount = int(d.Status.ReadyReplicas)
+	state.Target = target
+	state.InitialCount = int(current)
 	return nil, nil
 }
 
@@ -151,59 +146,40 @@ func (f PodCountCheckAction) Start(_ context.Context, _ *PodCountCheckState) (*a
 }
 
 func (f PodCountCheckAction) Status(_ context.Context, state *PodCountCheckState) (*action_kit_api.StatusResult, error) {
-	return statusPodCountCheckInternal(client.K8S, state), nil
-}
-
-func statusPodCountCheckInternal(k8s *client.Client, state *PodCountCheckState) *action_kit_api.StatusResult {
 	now := time.Now()
 
-	deployment := k8s.DeploymentByNamespaceAndName(state.Namespace, state.Deployment)
-	if deployment == nil {
-		return &action_kit_api.StatusResult{
-			Error: extutil.Ptr(action_kit_api.ActionKitError{
-				Title:  fmt.Sprintf("Deployment %s not found", state.Deployment),
-				Status: extutil.Ptr(action_kit_api.Errored),
-			}),
-		}
+	desired, current, err := f.GetDesiredAndCurrentPodCount(f.Client, state.Namespace, state.Target)
+	if err != nil {
+		return nil, err
 	}
 
-	readyCount := int(deployment.Status.ReadyReplicas)
-	desiredCount := 0
-	if deployment.Spec.Replicas != nil {
-		desiredCount = int(*deployment.Spec.Replicas)
-	} else if state.PodCountCheckMode == podCountEqualsDesiredCount || state.PodCountCheckMode == podCountLessThanDesiredCount {
-		return &action_kit_api.StatusResult{
-			Error: extutil.Ptr(action_kit_api.ActionKitError{
-				Title:  fmt.Sprintf("Deployment %s has no desired count.", state.Deployment),
-				Status: extutil.Ptr(action_kit_api.Errored),
-			}),
-		}
-	}
+	readyCount := int(current)
+	desiredCount := int(*desired)
 
 	var checkError *action_kit_api.ActionKitError
 	if state.PodCountCheckMode == podCountMin1 && readyCount < 1 {
 		checkError = extutil.Ptr(action_kit_api.ActionKitError{
-			Title:  fmt.Sprintf("%s has no ready pods.", state.Deployment),
+			Title:  fmt.Sprintf("%s has no ready pods.", state.Target),
 			Status: extutil.Ptr(action_kit_api.Failed),
 		})
 	} else if state.PodCountCheckMode == podCountEqualsDesiredCount && readyCount != desiredCount {
 		checkError = extutil.Ptr(action_kit_api.ActionKitError{
-			Title:  fmt.Sprintf("%s has only %d of desired %d pods ready.", state.Deployment, readyCount, desiredCount),
+			Title:  fmt.Sprintf("%s has only %d of desired %d pods ready.", state.Target, readyCount, desiredCount),
 			Status: extutil.Ptr(action_kit_api.Failed),
 		})
 	} else if state.PodCountCheckMode == podCountLessThanDesiredCount && readyCount == desiredCount {
 		checkError = extutil.Ptr(action_kit_api.ActionKitError{
-			Title:  fmt.Sprintf("%s has all %d desired pods ready.", state.Deployment, desiredCount),
+			Title:  fmt.Sprintf("%s has all %d desired pods ready.", state.Target, desiredCount),
 			Status: extutil.Ptr(action_kit_api.Failed),
 		})
 	} else if state.PodCountCheckMode == podCountIncreased && readyCount <= state.InitialCount {
 		checkError = extutil.Ptr(action_kit_api.ActionKitError{
-			Title:  fmt.Sprintf("%s's pod count didn't increase. Initial count: %d, current count: %d.", state.Deployment, state.InitialCount, readyCount),
+			Title:  fmt.Sprintf("%s's pod count didn't increase. Initial count: %d, current count: %d.", state.Target, state.InitialCount, readyCount),
 			Status: extutil.Ptr(action_kit_api.Failed),
 		})
 	} else if state.PodCountCheckMode == podCountDecreased && readyCount >= state.InitialCount {
 		checkError = extutil.Ptr(action_kit_api.ActionKitError{
-			Title:  fmt.Sprintf("%s's pod count didn't decrease. Initial count: %d, current count: %d.", state.Deployment, state.InitialCount, readyCount),
+			Title:  fmt.Sprintf("%s's pod count didn't decrease. Initial count: %d, current count: %d.", state.Target, state.InitialCount, readyCount),
 			Status: extutil.Ptr(action_kit_api.Failed),
 		})
 	}
@@ -212,11 +188,11 @@ func statusPodCountCheckInternal(k8s *client.Client, state *PodCountCheckState) 
 		return &action_kit_api.StatusResult{
 			Completed: true,
 			Error:     checkError,
-		}
+		}, nil
 	} else {
 		return &action_kit_api.StatusResult{
 			Completed: checkError == nil,
-		}
+		}, nil
 	}
 
 }