+++ extdeployment/common.go
@@ -4,10 +4,10 @@
 package extdeployment
 
 const (
-	DeploymentTargetType    = "com.steadybit.extension_kubernetes.kubernetes-deployment"
-	PodCountMetricActionId  = "com.steadybit.extension_kubernetes.pod_count_metric"
-	PodCountCheckActionId   = "com.steadybit.extension_kubernetes.pod_count_check"
-	RolloutRestartActionId  = "com.steadybit.extension_kubernetes.rollout-restart"
-	RolloutStatusActionId   = "com.steadybit.extension_kubernetes.rollout-status"
-	ScaleDeploymentActionId = "com.steadybit.extension_kubernetes.scale_deployment"
+	DeploymentTargetType            = "com.steadybit.extension_kubernetes.kubernetes-deployment"
+	PodCountMetricActionId          = "com.steadybit.extension_kubernetes.pod_count_metric"
+	DeploymentPodCountCheckActionId = "com.steadybit.extension_kubernetes.pod_count_check"
+	RolloutRestartActionId          = "com.steadybit.extension_kubernetes.rollout-restart"
+	RolloutStatusActionId           = "com.steadybit.extension_kubernetes.rollout-status"
+	ScaleDeploymentActionId         = "com.steadybit.extension_kubernetes.scale_deployment"
 )