+++ extstatefulset/common.go
@@ -4,6 +4,7 @@
 package extstatefulset
 
 const (
-	StatefulSetTargetType    = "com.steadybit.extension_kubernetes.kubernetes-statefulset"
-	ScaleStatefulSetActionId = "com.steadybit.extension_kubernetes.scale_statefulset"
+	StatefulSetTargetType            = "com.steadybit.extension_kubernetes.kubernetes-statefulset"
+	ScaleStatefulSetActionId         = "com.steadybit.extension_kubernetes.scale_statefulset"
+	StatefulSetPodCountCheckActionId = "com.steadybit.extension_kubernetes.pod_count_check_statefulset"
 )